moving attendance functions into separate class
